requirements.txt
PyYAML>=6.0
transformers>=4.44.0
torch>=2.2.0
accelerate>=0.33.0
tqdm>=4.66.0
